--- a/src/FresnelIntegrals.jl
+++ b/src/FresnelIntegrals.jl
@@ -1,27 +1,97 @@
 module FresnelIntegrals
 using SpecialFunctions
-export fresnelc
-export fresnels
-export fresnel
+using IrrationalConstants: sqrtπ
+export fresnelc, fresnels, fresnel
 
 """
-        fresnelc(z)
-Calculates the Fresnel cosine integral for the number z for
-    ``C(z) = \\int_{0}^{z} \\cos{\\left(\\frac{\\pi t^2}{2}\\right)}dt``
+    fresnelc(z::Number)
+
+Calculate the normalized Fresnel cosine integral
+```math
+C(z) = \\int_{0}^{z} \\cos{\\left(\\frac{\\pi t^2}{2}\\right)} \\, \\mathrm{d}t
+```
+for the number ``z``.
 """
-fresnelc(z::Number) = 0.25*(1-1im)*(1im*erf(0.5*(1-1im)*z*√(π)) + erf(0.5*(1+1im)*z*√(π)))
+function fresnelc(z::Number)
+    x = (z * sqrtπ) / 2
+    re_x, im_x = reim(x)
+    a = (re_x + im_x) + (im_x - re_x) * im
+    b = (re_x - im_x) + (im_x + re_x) * im
+    re_erf_a, im_erf_a = reim(erf(a))
+    re_erf_b, im_erf_b = reim(erf(b))
+    re_y = (re_erf_a - im_erf_a + re_erf_b + im_erf_b) / 4
+    im_y = (im_erf_a + re_erf_a - re_erf_b + im_erf_b) / 4
+    y = re_y + im_y * im
+    return y
+end
+function fresnelc(z::Real)
+    x = (z * sqrtπ) / 2
+    a = x + x * im
+    re_erf_a, im_erf_a = reim(erf(a))
+    y = (re_erf_a + im_erf_a) / 2
+    return y
+end
 
 """
-        fresnels(z)
-Calculates the Fresnel sine integral for the number z for
-    ``S(z) = \\int_{0}^{z} \\sin{\\left(\\frac{\\pi t^2}{2}\\right)}dt``
+    fresnels(z::Number)
+
+Calculate the normalized Fresnel sine integral
+```math
+S(z) = \\int_{0}^{z} \\sin{\\left(\\frac{\\pi t^2}{2}\\right)} \\, \\mathrm{d}t
+```
+for the number ``z``.
 """
-fresnels(z::Number) = 0.25*(1+1im)*(-1im*erf(0.5*(1-1im)*z*√(π)) + erf(0.5*(1+1im)*z*√(π)))
+function fresnels(z::Number)
+    x = (z * sqrtπ) / 2
+    re_x, im_x = reim(x)
+    a = (re_x + im_x) + (im_x - re_x) * im
+    b = (re_x - im_x) + (im_x + re_x) * im
+    re_erf_a, im_erf_a = reim(erf(a))
+    re_erf_b, im_erf_b = reim(erf(b))
+    re_y = (re_erf_a + im_erf_a + re_erf_b - im_erf_b) / 4
+    im_y = (im_erf_a - re_erf_a + re_erf_b + im_erf_b) / 4
+    y =  re_y + im_y * im
+    return y
+end
+function fresnels(z::Real)
+    x = (z * sqrtπ) / 2
+    a = x + x * im
+    re_erf_a, im_erf_a = reim(erf(a))
+    y = (re_erf_a - im_erf_a) / 2
+    return y
+end
 
 """
-        fresnel(z)
-Calculates the cosine and sine fresnel integrals
+    fresnel(z::Number)
+
+Calculate the normalized cosine and sine fresnel integrals.
+
+See also [`fresnels`](@ref), [`fresnelc`](@ref).
 """
-fresnel(z::Number) = (fresnelc(z),fresnels(z))
+function fresnel(z::Number)
+    x = (z * sqrtπ) / 2
+    re_x, im_x = reim(x)
+    a = (re_x + im_x) + (im_x - re_x) * im
+    b = (re_x - im_x) + (im_x + re_x) * im
+    re_erf_a, im_erf_a = reim(erf(a))
+    re_erf_b, im_erf_b = reim(erf(b))
+    re_y_sin = (re_erf_a + im_erf_a + re_erf_b - im_erf_b) / 4
+    im_y_sin = (im_erf_a - re_erf_a + re_erf_b + im_erf_b) / 4
+    re_y_cos = (re_erf_a - im_erf_a + re_erf_b + im_erf_b) / 4
+    im_y_cos = (im_erf_a + re_erf_a - re_erf_b + im_erf_b) / 4
+    y_sin = re_y_sin + im_y_sin * im
+    y_cos = re_y_cos + im_y_cos * im
+    return (y_cos, y_sin)
+end
+function fresnel(z::Real)
+    x = (z * sqrtπ) / 2
+    a = x + x * im
+    re_erf_a, im_erf_a = reim(erf(a))
+    y_sin = (re_erf_a - im_erf_a) / 2
+    y_cos = (re_erf_a + im_erf_a) / 2
+    return (y_cos, y_sin)
+end
+
+
 
 end # module