- Reformulated calculations to avoid multiplications
- Use of precise  √π provided by IrrationalConstants.jl
- Duplicates of functions for real inputs (faster calculations)
- Added tests for arbitrary values

All above was suggested in [https://github.com/JuliaMath/SpecialFunctions.jl/pull/407](https://github.com/JuliaMath/SpecialFunctions.jl/pull/407)

Tests are passing and the benchmark against QuadGK.jl is preserved