@devmotion 2022-08-12
> Duplicates of functions for real inputs (faster calculations)

They also return `Real` numbers now instead of `Complex` numbers with zero imaginary part.

On my computer, with the changes in this PR the benchmarks are roughly twice as fast for `Real` inputs:

## On master

```julia
julia> using FresnelIntegrals, QuadGK, BenchmarkTools

julia> @btime fresnelc(1.8)
  196.234 ns (0 allocations: 0 bytes)
0.3336329272215571 + 0.0im

julia> @btime quadgk(t->cos(π*t^2/2),0,1.8)
  950.158 ns (36 allocations: 1.02 KiB)
(0.33363292722155713, 1.348964606684433e-9)
```

## This PR

```julia
julia> using FresnelIntegrals, BenchmarkTools, QuadGK

julia> @btime fresnelc(1.8)
  95.555 ns (0 allocations: 0 bytes)
0.33363292722155713

julia> @btime quadgk(t->cos(π*t^2/2),0,1.8)
  988.727 ns (40 allocations: 1.08 KiB)
(0.33363292722155713, 1.348964606684433e-9)
```

Another improvement of this PR is that it avoids duplicate computations in `fresnel` (similar to how `sincos` does not actually call both `sin` and `cos` separately). That leads to a significant performance improvement (~ 4x):

## On master

```julia
julia> @btime fresnel(1.8)
  389.158 ns (0 allocations: 0 bytes)
(0.3336329272215571 + 0.0im, 0.4509387692675832 + 0.0im)
```

## This PR

```julia
julia> @btime fresnel(1.8)
  96.117 ns (0 allocations: 0 bytes)
(0.33363292722155713, 0.4509387692675829)
```

For completeness, here's a benchmark with complex inputs which only benefit from the improvements to `fresnel` (~ 2x faster):

## On master

```julia
julia> @btime fresnelc($(complex(1.8)))
  196.545 ns (0 allocations: 0 bytes)
0.3336329272215571 + 0.0im

julia> @btime fresnel($(complex(1.8)))
  389.356 ns (0 allocations: 0 bytes)
(0.3336329272215571 + 0.0im, 0.4509387692675832 + 0.0im)
```

## This PR

```julia
julia> @btime fresnelc($(complex(1.8)))
  197.625 ns (0 allocations: 0 bytes)
0.33363292722155713 + 1.3877787807814457e-17im

julia> @btime fresnel($(complex(1.8)))
  198.163 ns (0 allocations: 0 bytes)
(0.33363292722155713 + 1.3877787807814457e-17im, 0.4509387692675829 + 1.3877787807814457e-17im)
```