@emmt 2023-12-11
Compared to the base version, this PR automatically uses the correct numerical precision for all computations because there are no `Float64` constants.

Any idea why this has not yet been merged?