@kiranshila 2023-12-11
I'll update this for all the modern julia pacakging stuff too (haven't thought about this repo since I made it for my Antennas class 4 years ago!) :)