@kiranshila 2023-12-11
I'll also update the readme to include these new benchmarks